Move CompactionTask to compaction-core
