+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/task/CompactionTaskTerminateTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.execution;
+package sleeper.compaction.task;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
